requirements.txt
# To run: pip install -r requirements.txt
python-dotenv
numpy
pandas
scikit-learn
matplotlib
seaborn
pymongo
requests
pymongo[srv]==3.11
certifi
# -e .
